Move particlefile to _core
