+++ tests/test_particlefile.py
@@ -13,7 +13,7 @@ from parcels._core.utils.time import TimeInterval
 from parcels._datasets.structured.generic import datasets
 from parcels.kernels import AdvectionRK4
 from parcels.particle import Particle, create_particle_data, get_default_particle
-from parcels.particlefile import ParticleFile
+from parcels import ParticleFile
 from parcels.statuscodes import StatusCode
 from parcels import XGrid
 from tests.common_kernels import DoNothing
